01-tutorials/01-AgentCore-runtime/01-hosting-agent/01-strands-with-bedrock-model/requirements.txt
strands-agents
strands-agents-tools
uv
boto3>=1.39.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit